MediaWiki folder unpacked and deleted
Most things went into the “recycle” bin for separate inspection later
(dev stuff, potential tutorial topics to revisit, topics for forum,
maybe…) It will be for relevant parties to decide, but it’s not needed
in the base user docs. A temporary “help” folder was also created at
root, which will aide with filter FAQ stuff.